requirements.txt
streamlit>=1.32.0
plotly>=5.19.0
pandas>=2.0.0
groq>=0.9.0
python-dotenv>=1.0.0
